HR_View/requirements.txt
asgiref==3.5.2
defusedxml==0.7.1
Django==4.1.1
docopt==0.6.2
fonttools==4.37.3
fpdf2==2.5.7
num2words==0.5.12
Pillow==9.2.0
PyPDF2==2.10.9
sqlparse==0.4.3
svg.path==6.2
tzdata==2022.2
